Refactors CustomAlert, TranscriptForm
