+++ package.json
@@ -56,7 +56,6 @@
     "@babel/core": "^7.5.5",
     "@storybook/addon-a11y": "^5.0.8",
     "@storybook/addon-actions": "^5.1.9",
-    "@storybook/addon-knobs": "^5.0.8",
     "@storybook/addon-links": "^5.1.9",
     "@storybook/addon-viewport": "^5.0.8",
     "@storybook/addons": "^5.1.9",
