+++ packages/components/FormModal/index.js
@@ -1,4 +1,5 @@
 import React, { useState } from 'react';
+import { withKnobs, text, boolean, number } from '@storybook/addon-knobs';
 import PropTypes from 'prop-types';
 import Modal from 'react-bootstrap/Modal';
 import ItemForm from '../ItemForm/index.js';
@@ -8,21 +9,7 @@ const ItemFormModal = (props) => {
 
   const [ showModal, toggleShowModal ] = useState(props.showModal);
 
-  const form = () => {
-    if (props.type === 'transcript') {
-      return (
-        <TranscriptForm
-          { ...props }
-        />
-      );
-    } else {
-      return (
-        <ItemForm
-          { ...props }
-        />
-      );
-    }
-  };
+  const form = (props.type === 'transcript') ? <TranscriptForm { ...props }/> : <ItemForm { ...props }/>;
 
   return (
     <Modal show={ showModal } onHide={ () => toggleShowModal(!showModal) }>
@@ -30,7 +17,7 @@ const ItemFormModal = (props) => {
         <Modal.Title>{props.modalTitle}</Modal.Title>
       </Modal.Header>
       <Modal.Body>
-        {form()}
+        {form}
       </Modal.Body>
     </Modal>
   );
